+++ R/app_server.R
@@ -233,6 +233,13 @@ app_server <- function(input, output, session) {
                       app_version = golem::get_golem_options('app_version'),
                       metric_weights = metric_weights)
   
+  # Load server for the package dependencies tab.
+  dependencies_data <- packageDependenciesServer('packageDependencies',
+                                               selected_pkg,
+                                               user,
+                                               changes,
+                                               parent = session)
+  
   output$auth_output <- renderPrint({
     reactiveValuesToList(res_auth)
   })