+++ R/mod_pkg_explorer.R
@@ -49,6 +49,7 @@ mod_pkg_explorer_ui <- function(id){
 #' 
 #' @importFrom jsTreeR renderJstree jstree jstreeUpdate
 #' @importFrom shinyAce updateAceEditor
+#' @importFrom utils untar
 #'
 #' @noRd 
 mod_pkg_explorer_server <- function(id, selected_pkg,
@@ -96,7 +97,7 @@ mod_pkg_explorer_server <- function(id, selected_pkg,
       if (dir.exists(src_dir)) {
         pkgdir(src_dir)
       } else {
-        untar(file.path("tarballs", glue::glue("{selected_pkg$name()}_{selected_pkg$version()}.tar.gz")), exdir = "source")
+        utils::untar(file.path("tarballs", glue::glue("{selected_pkg$name()}_{selected_pkg$version()}.tar.gz")), exdir = "source")
         pkgdir(src_dir)
       }
     }) %>%