+++ tests/testthat/test-utils_startup.R
@@ -26,8 +26,8 @@ test_that("database creation", {
   expect_equal(names(pkg), c("id", "name", "version", "title", "description", "maintainer", "author", "license", "published_on", 
                              "score", "weighted_score", "decision_id", "decision_by", "decision_date", "date_added"))
   metric <- DBI::dbGetQuery(con, "SELECT * FROM metric")
-  expect_equal(nrow(metric), 12)
-  expect_equal(names(metric), c("id", "name", "long_name", "is_url", "is_perc", "description", "class", "weight"))
+  expect_equal(nrow(metric), 14) 
+  expect_equal(names(metric), c("id", "name", "long_name", "is_perc", "is_url", "description", "class", "weight"))
   pkg_metric <- DBI::dbGetQuery(con, "SELECT * FROM package_metrics")
   expect_equal(nrow(pkg_metric), 0)
   expect_equal(names(pkg_metric), c("id", "package_id", "metric_id", "value", "encode"))