+++ NEWS.md
@@ -8,6 +8,7 @@
 * Introduced cards to Database tab, reviewing contents of db (#587)
 * Prominently display the date a package was added (#486)
 * Fix issue where the repository being used to gather information was inconsistent
+* Added Package Dependencies page to "Package Metrics" tab, and two new cards to Maintenance Metrics page (#261)
 
 ### Squashed Bugs
 * Fixed busted button introduced with #547 (#592)